requirements.txt
streamlit
langchain_core
langchain_community
langchain_groq
langchain_ollama
langchain_huggingface
langchain_chroma
python-dotenv
pypdf